# goa/app/media_types.go
// Code generated by goagen v1.3.0, DO NOT EDIT.
//
// API "goa Practice": Application Media Types
//
// Command:
// $ goagen
// --design=github.com/imyslx/go-practice/goa/design
// --out=$(GOPATH)/src/github.com/imyslx/go-practice/goa
// --version=v1.3.0

package app

import (
	"github.com/goadesign/goa"
)

// Practice for goa API. (default view)
//
// Identifier: application/vnd.result+json; view=default
type Result struct {
	// message
	Message string `form:"message" json:"message" xml:"message"`
	// status
	Status int `form:"status" json:"status" xml:"status"`
}

// Validate validates the Result media type instance.
func (mt *Result) Validate() (err error) {

	if mt.Message == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "message"))
	}
	return
}